+++ src/threads.js
@@ -4704,7 +4704,14 @@ Process.prototype.doSwitchToScene = function (id, transmission) {
         message = this.inputOption(transmission.at(1)),
         ide, scenes, num, scene;
     this.assertAlive(rcvr);
-    this.assertType(message, ['text', 'number']);
+    this.assertType(message, ['text', 'number', 'list']);
+    if (message instanceof List && !message.canBeJSON()) {
+        // make sure only atomic leafs are inside the list
+        // don't actually encode the list as json, though
+        throw new Error(localize(
+            'can only send lists\nwith atomic data\nto another scene'
+        ));
+    }
     if (this.readyToTerminate) {
         // let the user press "stop" or "esc",
         // prevent "when this scene starts" hat blocks from directly